Constructing a factory now requires 5 units
Before this commit it only took one unit to construct a functioning
factory. This was quite overpowered, especially because it took ten
opposing units to then destroy the factory.

This commit makes it take 5 units to construct a functioning factory.
One unit to construct the outline of a factory and then 4 more units
colliding with that to get the factory to produce units.

It still takes 10 enemy units to destroy a factory. This might be too
many. As the AIs improved it may need rebalancing.