Stop unit cap being exceeded by multiple factories
The unit cap code was written before there was support for multiple
factories. It was adapted in a way that allowed going above the unit
cap when multiple factories produced units at once.

This code fixes the unit cap implementation to be much more solid. It
also improves the previous unit cap implementation by allowing units
to continue building when at the unit cap. This means that when a
player drops back below the unit cap they have units immediately ready
to go.