Produce turrets instead of vehicles
Before turrets are deployed they act as fully-functioning vehicles.
Constructing all vehicles as potential turrets allows quickly playing
with defensive/turtley AIs.

In the long run turrets, as with factories, will be more expensive and
will take a few units to construct. I'm not sure whether to have
an engineer vehicle which then constructs structures, or have a single
type of vehicle which then sacrifices itself to construct anything.

For small maps engineers feel too limited, but I'm nervous about the
work involved to allow bigger maps. Bigger maps already work and
render, but there's no way to zoom in. That would be a big refactor…