Zoom into the middle of what you can see
When I implemented zooming in 60a6bdcf3182b2d2898081ecef6ea991fed2c7f9
I left a known bug with zooming. When you zoomed in, you zoomed into
the top-left of the screen. This could be counteracted by moving the
camera with the arrow keys, so it wasn't critical to fix.

This commit fixes that issue. The camera position is now preserved
when zooming in and out.

Also, I think we're now at the point where a lot of this codebase can
be tested!