Move camera movement/zoom logic out of renderer
There is definite scope for improvement with the camera. Zooming is
too sudden and isn't based upon the mouse position; moving is at the
same speed regardless of zoom level.

Moving camera logic into a separate class will make it easier to
safely experiment with how to build a better camera.