Move Engineer Vehicle logic out of Bot unit
I think Bots should be hit-and-run units without construction
capabilities. Regardless of the details of that, I'll definitely want
other units to be able to construct structures. This commit moves that
logic out of Bot into an Engineer class.

This definitely isn't perfect. It's unusual compared to making
everything else a capability mixin. But it's good enough for now. :)