Use quadtrees to accelerate collision detection
Existing vehicles mostly cause damage by colliding with objects.
That makes detecting collisions a key part of the game. This commit
replaces the bruteforce, quadratic collision detection with a smarter
approach based upon quadtrees.

When experimenting I saw the new collision detection taking ~15% of
the computation time of the old, and scaling much better to larger
numbers of units. To the eye it appears to have worked out well.