Add strategic icons to units
This commit gives all units a strategic icon. This is only visible
when zoomed out so far that the underlying units are almost invisible.
Its purpose is to support more strategic games on much larger maps,
where without the icons it would be impossible to manually see units.

The implementation of icon is a little weird and could do with a
refactor. Since the shapes being used aren't being scaled, it ought to
be simpler to use Ruby2D's `Square` and `Text` classes rather than my
camera-scaled ones?