Workaround icon text exhausting file descriptors
The previous commit added a strategic icon to each unit. Each includes
a one-character piece of text, so for the first time there are a
significant number of Ruby2D `Text` classes [1] in use.

Ruby2D is a wrapper around Simple2D. It turns out that the Simple2D
library hardcodes allocating one file description per piece of text
[2]. That means I can't cache font files and control file descriptor
usage without patching Simple2D, which would involve a bit of work.

This commit implements a trivial workaround: increase the RTS game's
file descriptor limit to be at least as high as we could possibly
need. It's a bit horrible but it works for any number of units that
have decent performance right now.

[1] https://www.ruby2d.com/learn/text/
[2] https://github.com/simple2d/simple2d/blob/1060a6efccfb8e92745082fcc57d4cac2ad55897/src/text.c#L28-L34