Use Ruby2D::Window.objects rather than duplicating
This commit is a bit ugly, but it stops Renderer from having to keep
track of what shapes are being rendered in order to resize them when
zooming and moving the camera.

This commit makes Renderer (ab)use the way that Ruby2D::Window already
keeps track of the shapes being rendered. There are a few ugly hacks
used to make this work and it will probably be more sensible in the
long run to patch `Ruby2D::Window`'s `add` and `remove` methods to
track state in a way that helps us more.

This does cause a performance regression. Composite shapes with custom
rescaling logic, such as Icon, will be rescaled once for every child
shape they have. This isn't a big deal as camera movement happens a
lot less than rendering frames.