Power Sources have a structure but cannot be owned
This commit changes Power Sources so that they can be home to a
structure but cannot be directly owned by a player. This is a step
towards being able to build Generator structures on top of Power
Source locations.

Game logic around capturing Power Sources is temporarily commented
out. It will be adapted and reintroduced once Generators can be built
atop Power Sources.